[3/3] OmniGears: move UI Tuner to dev settings
-add direct link to quick settings
-add direct link to icons blacklist
-split network traffic into separate fragment
