+++ res/values/custom_strings.xml
@@ -223,4 +223,8 @@
     <string name="status_bar_network_activity_title">Activity indicators</string>
     <string name="status_bar_network_activity_summary">Show indicators in signal bars</string>
 
+    <string name="quick_settings_title">Quick Settings</string>
+    <string name="quick_settings_summary">Configure quick settings tiles</string>
+    <string name="status_bar_icon_blacklist_title">Status bar icons</string>
+    <string name="status_bar_icon_blacklist_summary">Configure what icons are shown on status bar</string>
 </resources>
