+++ res/xml/bars_settings.xml
@@ -33,57 +33,20 @@
                 android:title="@string/battery_title"
                 android:fragment="org.omnirom.omnigears.interfacesettings.StatusbarBatterySettings" />-->
 
+            <PreferenceScreen
+                    android:title="@string/status_bar_icon_blacklist_title"
+                    android:summary="@string/status_bar_icon_blacklist_summary" >
+                <intent android:action="android.intent.action.MAIN"
+                        android:targetPackage="com.android.systemui"
+                        android:targetClass="com.android.systemui.tuner.StatusBarActivity" />
+            </PreferenceScreen>
         </PreferenceCategory>
 
         <!-- Network traffic meter -->
-        <PreferenceCategory
+        <PreferenceScreen
             android:key="category_network_traffic"
-            android:title="@string/network_traffic_title">
-
-            <!--<com.android.settings.preference.SystemCheckBoxPreference
-                android:key="status_bar_network_activity"
-                android:title="@string/status_bar_network_activity_title"
-                android:summary="@string/status_bar_network_activity_summary"
-                android:defaultValue="false"/>-->
-
-            <ListPreference
-                android:key="network_traffic_state"
-                android:title="@string/show_network_traffic_state"
-                android:entries="@array/show_network_traffic_state_entries"
-                android:entryValues="@array/show_network_traffic_state_values"
-                android:persistent="false" />
-
-            <ListPreference
-                android:key="network_traffic_unit"
-                android:title="@string/show_network_traffic_unit"
-                android:entries="@array/show_network_traffic_unit_entries"
-                android:entryValues="@array/show_network_traffic_unit_values"
-                android:persistent="false" />
-
-            <ListPreference
-                android:key="network_traffic_period"
-                android:title="@string/show_network_traffic_frequency"
-                android:entries="@array/show_network_traffic_frequency_entries"
-                android:entryValues="@array/show_network_traffic_frequency_values"
-                android:persistent="false" />
-
-            <com.android.settings.preference.SystemCheckBoxPreference
-                android:key="network_traffic_autohide"
-                android:title="@string/network_traffic_autohide_title"
-                android:summary="@string/network_traffic_autohide_summary"
-                android:defaultValue="false"/>
-
-            <com.android.settings.preference.SeekBarPreference
-                android:key="network_traffic_autohide_threshold"
-                android:title="@string/network_traffic_autohide_threshold"
-                android:max="100"
-                settings:min="0"
-                settings:unitsLeft=""
-                settings:unitsRight="kB/s"
-                android:dependency="network_traffic_autohide"
-                android:persistent="false" />
-
-        </PreferenceCategory>
+            android:title="@string/network_traffic_title"
+            android:fragment="org.omnirom.omnigears.interfacesettings.NetworkTraffic" />
 
         <PreferenceCategory
             android:key="notification_category"
@@ -94,18 +57,13 @@
                     android:title="@string/status_bar_custom_header_title"
                     android:summary="@string/status_bar_custom_header_summary"
                     android:defaultValue="false"/>
+
+            <PreferenceScreen
+                    android:title="@string/quick_settings_title"
+                    android:summary="@string/quick_settings_summary" >
+                <intent android:action="android.intent.action.MAIN"
+                        android:targetPackage="com.android.systemui"
+                        android:targetClass="com.android.systemui.tuner.QsTunerActivity" />
+            </PreferenceScreen>
         </PreferenceCategory>
-
-        <!--<PreferenceCategory
-            android:key="navbar_category"
-            android:title="@string/navbar_title" >
-
-            <ListPreference
-                android:key="navbar_recents_style"
-                android:title="@string/navbar_recents_style_title"
-                android:entries="@array/navbar_recents_style_entries"
-                android:entryValues="@array/navbar_recents_style_values"
-                android:persistent="false"/>
-        </PreferenceCategory>-->
-
 </PreferenceScreen>
